+++ packages/checkout/components/index.js
@@ -1,0 +1,8 @@
+export * from './totals';
+export { default as TotalsWrapper } from './totals-wrapper';
+export { default as ExperimentalOrderMeta } from './order-meta';
+export { default as ExperimentalDiscountsMeta } from './discounts-meta';
+export { default as ExperimentalOrderShippingPackages } from './order-shipping-packages';
+export { default as Panel } from './panel';
+export { default as Button } from './button';
+export { default as Label } from './label';
