+++ packages/checkout/components/label/index.ts
