+++ packages/checkout/components/panel/index.tsx
