+++ packages/checkout/components/panel/style.scss
