+++ packages/checkout/components/totals-wrapper/style.scss
