+++ packages/checkout/components/totals/item/style.scss
