+++ packages/checkout/components/totals/taxes/style.scss
