+++ packages/checkout/index.js
@@ -1,13 +1,6 @@
-export * from './totals';
+export * from './components';
 export * from './utils';
 export * from './slot';
 export * from './registry';
-export { default as TotalsWrapper } from './wrapper';
 export * from './blocks-registry';
-export { default as ExperimentalOrderMeta } from './order-meta';
-export { default as ExperimentalDiscountsMeta } from './discounts-meta';
-export { default as ExperimentalOrderShippingPackages } from './order-shipping-packages';
-export { default as Panel } from './panel';
 export { SlotFillProvider } from 'wordpress-components';
-export { default as Button } from './button';
-export { default as Label } from './label';
